% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append<-.list}
\alias{append<-.list}
\alias{append<-}
\title{Assignment function to append values to existing list}
\usage{
\method{append}{list}(x, name = NULL) <- value
}
\arguments{
\item{x}{existing list}

\item{name}{optional name of new element}

\item{value}{new value to add to existing list}
}
\description{
Assignment function to append values to existing list
}
\examples{
x <- list()
append(x) <- 1
append(x, name = "a") <- 2
# duplicated names are allowed
append(x, name = "a") <- 3
x
}
