% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{bisection}
\alias{bisection}
\title{Root finding by bisection}
\usage{
bisection(f, interval, niter = 6, tol = 1e-12, verbose = TRUE, ...)
}
\arguments{
\item{f}{function to find root of (monotonic)}

\item{interval}{a vector containing the end-points of the interval to be
searched for the root}

\item{niter}{number of iterations}

\item{tol}{stopping criterion (absolute difference in function evaluated at
end points of current interval)}

\item{verbose}{if TRUE additional messages are printed throughout the
optimization}

\item{...}{additional arguments passed to \code{f}}
}
\value{
numeric specifying the root
}
\description{
Root finding by bisection
}
\examples{
testf <- function(x) (x - sqrt(2))
carts:::bisection(testf, c(0, 1000), niter = 30)
}
