% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar_sim.R
\name{covar_bootstrap}
\alias{covar_bootstrap}
\title{Sample from empirical distribution of covariate data}
\usage{
covar_bootstrap(data, subset = NULL)
}
\arguments{
\item{data}{data.frame}

\item{subset}{optional columns to select from data frame}
}
\value{
random generator (function)
}
\description{
Sample from empirical distribution of covariate data
}
