% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar_sim.R
\name{derive_covar_distribution}
\alias{derive_covar_distribution}
\title{Derive covariate distribution from covariate data type}
\usage{
derive_covar_distribution(covar)
}
\arguments{
\item{covar}{Vector with covariates}
}
\value{
lava package random generator function
}
\description{
Derive covariate distribution (for outcome regression) for
integers (Poisson), numeric (Gaussian) and binary (Binomial) data. Raise an
error for other data types.
}
\author{
Benedikt Sommer
}
