% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{est_glm}
\alias{est_glm}
\alias{est_gee}
\alias{est_geebin}
\alias{est_glmbin}
\title{Construct estimator for the treatment effect in RCT}
\usage{
est_glm(
  response = "y",
  treatment = "a",
  covariates = NULL,
  offset = NULL,
  id = NULL,
  level = 0.95,
  family = gaussian(),
  target.parameter = treatment,
  ...
)
}
\arguments{
\item{response}{(character) Response variable}

\item{treatment}{(character) Treatment variable}

\item{covariates}{(character; optional) Single or vector of covariates}

\item{offset}{(character; optional) Model offset}

\item{id}{(character; optional) Subject id variable}

\item{level}{(numeric) Confidence interval level}

\item{family}{(family or character) Exponential family that is supported by
\link[stats:glm]{stats::glm} and \link[MASS:glm.nb]{MASS::glm.nb}}

\item{target.parameter}{(character) Target parameter from model output}

\item{...}{Additional arguments to \link[lava:estimate.default]{lava::estimate}}
}
\value{
function
}
\description{
Returns an estimator for the estimation of a treatment effect
with parametric models (\link[stats:glm]{stats::glm} and \link[MASS:glm.nb]{MASS::glm.nb}). The returned
estimator is a function with a single argument (data) and returns the
treatment effect estimate, which is estimated with \link[lava:estimate.default]{lava::estimate}
}
\examples{
trial <- Trial$new(
    covariates = function(n) data.frame(a = rbinom(n, 1, 0.5), x = rnorm(n)),
    outcome = setargs(outcome_count,
      mean = ~ 1 + a*x,
      par = c(1, -0.1, 0.5, 0.2),
      overdispersion = 2)
)
dd <- trial$simulate(3e2)

# crude mean comparison between arms (default behavior; y ~ a)
est <- est_glm(family = poisson)
est(dd)

# linear adjustment with one covariate (y ~ a + x)
est <- est_glm(family = poisson, covariates = "x")
est(dd)

# return estimates of all linear coefficients (useful for debugging)
est <- est_glm(family = poisson, covariates = "x", target.parameter = NULL)
est(dd)

# comparing robust and non-robust standard errors of poisson estimator by
# passing robust argument via ... to lava::estimate
estimators <- list(
  robust = est_glm(family = poisson),
  non.robust = est_glm(family = poisson, robust = FALSE)
)
res <- do.call(rbind, lapply(estimators, \(est) est(dd)$coefmat))
rownames(res) <- names(estimators)
res
}
\seealso{
\link{Trial}
}
\author{
Klaus Kähler Holst
}
