% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{est_phreg}
\alias{est_phreg}
\title{Marginal Cox proportional hazards model for the treatment effect in
RCT}
\usage{
est_phreg(
  response = "Surv(time, status)",
  treatment = "a",
  level = 0.95,
  id = NULL
)
}
\arguments{
\item{response}{Response variable (character or formula). Default:
"Surv(time, status)"}

\item{treatment}{Treatment variable (character)}

\item{level}{Confidence interval level}

\item{id}{Optional subject id variable (character)}
}
\value{
function
}
\description{
Marginal Cox proportional hazards model for the treatment effect in
RCT
}
\seealso{
\link{Trial} \link{est_adj}
}
\author{
Klaus Kähler Holst
}
