% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar_sim.R
\name{get_factor_levels}
\alias{get_factor_levels}
\title{Get levels for factor columns in data.table}
\usage{
get_factor_levels(data)
}
\arguments{
\item{data}{Covariate \code{data.table}}
}
\description{
Get levels for factor columns in data.table
}
