% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_models.R
\name{outcome_continuous}
\alias{outcome_continuous}
\title{Simulate from continuous outcome model given covariates}
\usage{
outcome_continuous(
  data,
  mean = NULL,
  par = NULL,
  sd = 1,
  het = 0,
  outcome.name = "y",
  remove = c("id", "num"),
  family = gaussian(),
  ...
)
}
\arguments{
\item{data}{covariate matrix}

\item{mean}{formula specifying design from 'data' or a function that maps x
to the mean value. If NULL all main-effects of the covariates will be used}

\item{par}{Regression coefficients (default zero). Can be given as a named
list corresponding to the column names of \code{model.matrix}}

\item{sd}{standard deviation of measurement error}

\item{het}{Introduce variance hetereogeneity by adding a residual term
\eqn{het \cdot \mu_x \cdot e}, where \eqn{\mu_x} is the mean given
covariates and \eqn{e} is an independent standard normal distributed
variable. This term is in addition to the measurement error introduced by
the \code{sd} argument.}

\item{outcome.name}{Name of outcome variable ("y")}

\item{remove}{variables that will be removed from input x (if formula is not
specified)}

\item{family}{exponential family (default \code{gaussian(identity)})}

\item{...}{Additional arguments passed to lower level functions}
}
\value{
data.table
}
\description{
Simulate from continuous outcome model with mean
\deqn{g(\text{par}^\top X)} where \eqn{X} is the design matrix specified by
the formula, and \eqn{g} is the link function specified by the family
argument
}
\seealso{
\link{outcome_count} \link{outcome_binary} \link{outcome_lp}
}
