% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_models.R
\name{outcome_count}
\alias{outcome_count}
\title{Simulate from count model given covariates}
\usage{
outcome_count(
  data,
  mean = NULL,
  par = NULL,
  outcome.name = "y",
  exposure = 1,
  remove = c("id", "num"),
  zero.inflation = NULL,
  overdispersion = NULL,
  ...
)
}
\arguments{
\item{data}{covariate matrix}

\item{mean}{formula specifying design from 'x' or a function that maps x to
the mean value. The response variable needs to be left undefined, i.e. \code{~ x1 + x2} defines the mean on the link-scale as a linear function of
covariates x1 and x2 (see examples). If NULL all main-effects of the
covariates will be used.}

\item{par}{Regression coefficients (default zero). Can be given as a named
list corresponding to the column names of \code{model.matrix}}

\item{outcome.name}{Name of outcome variable ("y")}

\item{exposure}{Exposure times. Either a scalar, vector or function.}

\item{remove}{Variables that will be removed from input x (if formula is not
specified)}

\item{zero.inflation}{vector of probabilities or a function of the covariates
'x' including an extra column 'rate' with the rate parameter.}

\item{overdispersion}{variance of gamma-frailty either given as a numeric
vector or a function of the covariates 'x' with an extra column 'rate'
holding the rate parameter 'rate'}

\item{...}{Additional arguments passed to lower level functions}
}
\value{
data.table with added exposure column
}
\description{
Simulate from count model with intensity \deqn{\lambda =
  \text{exposure-time}\exp(\text{par}^\top X)} where \eqn{X} is the design
matrix specified by the formula
}
\examples{
set.seed(24)
covariates <- function(n) data.frame(a = rbinom(n, 1, 0.5))
outcome <- setargs(
  outcome_count,
  mean = ~ 1 + a,
  par = log(c(2.5, 0.65)),
  overdispersion = 1 / 2,
  exposure = 2 # identical exposure time for all subjects
)
trial <- Trial$new(covariates = covariates, outcome = outcome)
trial$simulate(5)
# alternating exposure times between subjects
trial$args_model(exposure = c(1, 2))
trial$simulate(5)
# treatment-dependent exposure times
trial$args_model(exposure = function(dd) 1 - 0.5 * dd$a)
trial$simulate(5)
}
\seealso{
\link{outcome_binary} \link{outcome_continuous} \link{outcome_lp}
}
