% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_models.R
\name{outcome_recurrent}
\alias{outcome_recurrent}
\title{UNDER DEVELOPMENT: Outcome model for recurrent events with terminal
events end-points}
\usage{
outcome_recurrent(
  data,
  lp = NULL,
  par = NULL,
  outcome.name = c("time", "status"),
  remove = c("id", "num"),
  model = NULL,
  death.model = NULL,
  death.lp = NULL,
  death.par = NULL,
  cens.model = NULL,
  cens.lp = NULL,
  cens.par = NULL,
  ...
)
}
\arguments{
\item{data}{data.frame (covariates)}

\item{lp}{linear predictor (formula or function)}

\item{par}{optional list of model parameter}

\item{outcome.name}{names of outcome (time and censoring status)}

\item{remove}{variables that will be removed from input data (if formula is
not specified)}

\item{model}{optional \link[mets:phreg]{mets::phreg} object}

\item{death.model}{optional model for death (terminal) events}

\item{death.lp}{optional death linear predictor argument (formula or
function)}

\item{death.par}{optional list of death model parameters}

\item{cens.model}{optional model for censoring mechanism}

\item{cens.lp}{optional censoring linear predictor argument
(formula or function)}

\item{cens.par}{optional list of censoring model parameters}

\item{...}{Additional arguments to \link{outcome_lp}}
}
\value{
function (random generator)
}
\description{
This function is still in an experimental state where the
interface and functionality might change in the future
}
\seealso{
\link{outcome_count} \link{outcome_lp} \link{outcome_binary} \link{outcome_continuous}
\link{outcome_phreg}
}
