% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_models.R
\name{outcome_shared}
\alias{outcome_shared}
\title{Outcome model}
\arguments{
\item{data}{(data.table) Covariate data, usually the output of the covariate
model of a \link{Trial} object.}

\item{par}{(numeric) Regression coefficients (default zero). Can be given as
a named list corresponding to the column names of \code{model.matrix}}

\item{outcome.name}{Name of outcome variable ("y")}

\item{formula}{(formula) Formula specifying response and design matrix.}

\item{remove}{Variables that will be removed from input \code{data} (if formula is
not specified).}

\item{lp}{formula specifying design from 'data' or a function that maps x
to the conditional mean value on the link scale (see examples). If NULL all
main-effects of the covariates will be used.}

\item{...}{Additional arguments passed to \code{mean} function (see examples)}
}
\value{
data.table
}
\description{
Outcome model
}
