% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_run.R
\name{summary.trial.estimates}
\alias{summary.trial.estimates}
\title{Summary method for simulated trials (trial.estimates objects)}
\usage{
\method{summary}{trial.estimates}(
  object,
  level = 0.05,
  null = 0,
  ni.margin = NULL,
  alternative = c("!=", "<", ">"),
  reject.function = NULL,
  true.value = NULL,
  nominal.coverage = 0.9,
  ...
)
}
\arguments{
\item{object}{(trial.estimates)}

\item{level}{significance level}

\item{null}{null hypothesis to test}

\item{ni.margin}{non-inferiority margin}

\item{alternative}{alternative hypothesis (not equal !=, less <, greater >)}

\item{reject.function}{Optional function calculating whether to reject the
null hypothesis. Should be a function of one or more of the following
arguments: estimate (parameter estimate), stderr (Standard error of the
estimate), lower (lower confidence limit), upper (upper confidence limit)}

\item{true.value}{Optional true parameter value. When provided the root-mean
square error (RMSE) and coverage of the estimated confidence limits will
also be calculated.}

\item{nominal.coverage}{Width of confidence limits used to calculate the
empirical coverage around \code{true.value}. The default value calculates the
coverage for the central 90\% confidence interval.}

\item{...}{additional arguments to lower level functions}
}
\value{
matrix with results of each estimator stored in separate rows
}
\description{
Summary method for the Monte Carlo studies of different
estimators for the treatment effect in a randomized clinical trial
(trial.estimates objects). The power of both a superiority test (one-sided
or two-sided) and a non-inferiority test are reported, together with the
summary statistics of the different estimators.
}
