% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carts-package.R
\docType{package}
\name{carts-package}
\alias{carts-package}
\alias{carts}
\title{carts: Simulation-based assessment of Covariate Adjustment in Randomized Trials}
\description{
Monte Carlo simulation framework for different randomized clinical trial designs with a special emphasis on estimators based on covariate adjustment.

Monte Carlo simulation framework for different randomized
clinical trial designs with a special emphasis on estimators based on
covariate adjustment.
}
\examples{
rnb(10, 1, 1)
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/NN-AI-Analytics/carts/issues}
}

}
\author{
\strong{Maintainer}: Benedikt Sommer \email{benediktsommer92@gmail.com}

Authors:
\itemize{
  \item Klaus K. Holst \email{klaus@holst.it}
  \item Foroogh Shamsi \email{foroogh.shamsi1@gmail.com}
}


Benedikt Sommer, Klaus Holst, Foroogh Shamsi
}
\keyword{internal}
\keyword{package}
