% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carts-package.R
\docType{package}
\name{carts-package}
\alias{carts-package}
\alias{carts}
\title{carts: Simulation-Based Assessment of Covariate Adjustment in Randomized Trials}
\description{
Monte Carlo simulation framework for different randomized clinical trial designs with a special emphasis on estimators based on covariate adjustment.

Monte Carlo simulation framework for different randomized
clinical trial designs with a special emphasis on estimators based on
covariate adjustment.
}
\examples{
\dontrun{
trial <- Trial$new(
  covariates = \(n) data.frame(a = rbinom(n, 1, 0.5), x = rnorm(n)),
  outcome = setargs(outcome_count, par = c(1, 0.5, 1), overdispersion = 0.7)
)

trial$estimators(
  unadjusted = est_glm(family = "poisson"),
  adjusted = est_glm(family = "poisson", covariates = "x")
)

trial$run(n = 200, R = 100)
trial$summary()
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://novonordisk-opensource.github.io/carts/}
  \item \url{https://github.com/NovoNordisk-OpenSource/carts}
  \item Report bugs at \url{https://github.com/NovoNordisk-OpenSource/carts/issues}
}

}
\author{
\strong{Maintainer}: Benedikt Sommer \email{benediktsommer92@gmail.com}

Authors:
\itemize{
  \item Klaus K. Holst \email{klaus@holst.it}
  \item Foroogh Shamsi \email{foroogh.shamsi1@gmail.com}
}


Benedikt Sommer, Klaus Holst, Foroogh Shamsi
}
\keyword{internal}
\keyword{package}
