% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{est_adj}
\alias{est_adj}
\title{Construct estimator for the treatment effect in RCT based on covariate
adjustment}
\usage{
est_adj(
  response = "y",
  treatment = "a",
  covariates = NULL,
  offset = NULL,
  id = NULL,
  family = gaussian(),
  level = 0.95,
  treatment.effect = NULL,
  nfolds = 1,
  ...
)
}
\arguments{
\item{response}{Response variable (character)}

\item{treatment}{Treatment variable (character)}

\item{covariates}{Optional covariates (character)}

\item{offset}{Optional offset (character)}

\item{id}{Optional subject id variable (character)}

\item{family}{Exponential family (default gaussian)}

\item{level}{Confidence interval level}

\item{treatment.effect}{(optional) function describing treatment effect given
mean potential outcomes. Default is the average treatment effect
(difference in expected outcome, (x,y) -> (y-x)) unless the family is
binomial or poisson, in which case the treatment effect is the difference
in log expected outcomes (x,y) -> log(y/x).}

\item{nfolds}{Number of folds for estimating the conditional average
treatment effect with double machine learning.}

\item{...}{Additional arguments to \link[targeted:learner_glm]{targeted::learner_glm}}
}
\value{
function
}
\description{
Efficient estimator of the treatment effect based on the
efficient influence function. This involves a model for the conditional
mean of the outcome variable given covariates.
}
\details{
...
}
\seealso{
\link{Trial} \link{est_glm}
}
\author{
Klaus Kähler Holst
}
