% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_models.R
\name{outcome_binary}
\alias{outcome_binary}
\title{Simulate from binary model given covariates}
\usage{
outcome_binary(
  data,
  mean = NULL,
  par = NULL,
  outcome.name = "y",
  remove = c("id", "num"),
  family = binomial(logit),
  ...
)
}
\arguments{
\item{data}{covariate matrix}

\item{mean}{formula specifying design from 'data' or a function that maps x
to the mean value. If NULL all main-effects of the covariates will be used}

\item{par}{Regression coefficients (default zero). Can be given as a named
list corresponding to the column names of \code{model.matrix}}

\item{outcome.name}{Name of outcome variable ("y")}

\item{remove}{variables that will be removed from input x (if formula is not
specified)}

\item{family}{exponential family (default \code{binomial(logit)})}

\item{...}{Additional arguments passed to lower level functions}
}
\value{
data.table
}
\description{
Simulate from binary model with probability \deqn{\pi =
  g(\text{par}^\top X)} where \eqn{X} is the design matrix specified by the
formula, and \eqn{g} is the link function specified by the family argument
}
\seealso{
\link{outcome_count} \link{outcome_lp} \link{outcome_continuous}
}
