% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_models.R
\name{outcome_phreg}
\alias{outcome_phreg}
\title{Outcome model for time-to-event end-points (proportional hazards)}
\usage{
outcome_phreg(
  data,
  lp = NULL,
  par = NULL,
  outcome.name = c("time", "status"),
  remove = c("id", "num"),
  model = NULL,
  cens.model = NULL,
  cens.lp = NULL,
  cens.par = NULL,
  ...
)
}
\arguments{
\item{data}{(data.table) Covariate data, usually the output of the covariate
model of a \link{Trial} object.}

\item{lp}{linear predictor (formula or function)}

\item{par}{optional list of model parameter}

\item{outcome.name}{names of outcome (time and censoring status)}

\item{remove}{Variables that will be removed from input \code{data} (if formula is
not specified).}

\item{model}{optional \link[mets:phreg]{mets::phreg} object}

\item{cens.model}{optional model for censoring mechanism}

\item{cens.lp}{censoring linear predictor argument (formula or function)}

\item{cens.par}{list of censoring model parameters}

\item{...}{Additional arguments to \link{outcome_lp}}
}
\value{
function (random generator)
}
\description{
Outcome model for time-to-event end-points (proportional hazards)
}
\examples{
\dontrun{
library("survival")
data(pbc, package = "survival")
pbc0 <- na.omit(pbc) |>
  transform(trt = factor(trt, labels = c("Active", "Placebo")) |>
    relevel(ref = "Placebo"))

fit1 <- mets::phreg(Surv(time, status > 0) ~ age + sex * trt, data = pbc0)

covar <- covar_bootstrap(pbc0, subset = c("age", "sex")) \%join\%
  function(n, ...) data.frame(trt = sample(pbc0$trt, n, replace = TRUE))

outcome <- setargs(
  outcome_phreg,
  model = fit1,
  par = list("trtActive" = 0)
)

xx <- covar(5e3)
pbc1 <- outcome(xx) |> cbind(xx)
mets::phreg(formula(fit1), data = pbc1)


## Introducing additional interactions
fit2 <- mets::phreg(Surv(time, status > 0) ~ age + sex + trt, data = pbc0)

outcome <- setallargs(
  outcome_phreg,
  model = fit2,
  par = list("trtActive" = -.5, "age:trtActive" = 0),
  treatment = "trt",
  default.parameter = -0.2
)

xx <- covar(1e4)
attr(outcome(xx), "par")
pbc1 <- outcome(xx) |> cbind(xx)
mets::phreg(Surv(time, status) ~ (age + sex) * trt, pbc1)
rm(pbc1, xx)
}
}
\seealso{
\link{outcome_count} \link{outcome_lp} \link{outcome_binary} \link{outcome_continuous}
}
\author{
Klaus Kähler Holst
}
